Refactor the training pipeline so it can be reused
